requirements.txt
python-dotenv
fastapi[standard]
pydantic
SQLAlchemy
psycopg2-binary
pymysql
axios
python-dateutil
boto3
streamlit
pyjwt
passlib[bcrypt]
pytest-asyncio